requirements.txt
selenium
pymongo
python-dotenv
webdriver-manager
flask